Used floats instead of doubles, removed the artifact, commented code and bug fixed objective value

+++ Simplex Instance Generator/src/InstanceGenerator.java
@@ -7,34 +7,34 @@ public class InstanceGenerator {
     public SimplexInstance generate(int variableNumber, int constraintNumber){
         SimplexInstance instance = initializeFormulation(variableNumber, constraintNumber);
 
-        double expectedObjective = simplexSolver.solveSimplex(instance);
+        float expectedObjective = simplexSolver.solveSimplex(instance);
         instance.setExpectedObjective(expectedObjective);
 
         return instance;
     }
 
     private SimplexInstance initializeFormulation(int variableNumber, int constraintNumber) {
-        double[][] constraints = new double[constraintNumber][variableNumber];
-        double[] initialCoefficients = new double[variableNumber];
-        double[] initialConstants = new double[constraintNumber];
+        float[][] constraints = new float[constraintNumber][variableNumber];
+        float[] initialCoefficients = new float[variableNumber];
+        float[] initialConstants = new float[constraintNumber];
 
         // initialize the initial objective value.
-        double initialObjective = random.nextDouble()*100;
+        float initialObjective = random.nextFloat() * 100;
 
         // Initialize coefficients of the objective function
         for (int i = 0; i < variableNumber; i++) {
-            initialCoefficients[i] = random.nextDouble() * 100;
+            initialCoefficients[i] = random.nextFloat() * 100;
         }
 
         // initialize constants of the constraints
         for (int i = 0; i < constraintNumber; i++) {
-            initialConstants[i] = random.nextDouble() * 100;
+            initialConstants[i] = random.nextFloat() * 100;
         }
 
         // initialize the coefficients of the constraints
         for (int i = 0; i < constraintNumber; i++) {
             for (int j = 0; j < variableNumber; j++) {
-                constraints[i][j] = random.nextDouble() * 100;
+                constraints[i][j] = random.nextFloat() * 100;
             }
         }
